Create data processing folder structure
Create a structure for the folder "data-processing" that mimicks the structure of folder "data". Each script in "data-processing" is to be placed in the folder that corresponds to the one it extracts files to in "data".
Eg. extract_json_greece.py reads files from "/data/all_countries/JohnsHopkinsCSSE/" and extracts a file to "/data/greece/JohnsHopkinsCSSE/". Therefore, it should be placed in "data-processing/greece/JohnsHopkinsCSSE/".